requirements.txt
Django>=5.0
 djangorestframework>=3.14
 djangorestframework-api-key>=3.0 
